crojas-azurian-az-track-445649e83e2f[DANIEL]/requirements.txt
amqp==2.2.1
anyjson==0.3.3
appdirs==1.4.3
billiard==3.5.0.3
boto==2.48.0
celery==4.1.0
Django==1.11.4
django-celery-beat==1.0.1
django-filter==1.0.4
django-storages==1.6.5
djangorestframework==3.6.4
et-xmlfile==1.0.1
gunicorn==19.7.1
httplib2==0.10.3
jdcal==1.3
kombu==4.1.0
Markdown==2.6.9
mysql-python==1.2.5
newrelic==2.90.0.75
numpy==1.13.1
odfpy==1.3.5
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
protobuf==3.4.0
pyasn1==0.3.3
pyasn1-modules==0.1.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2017.2
PyYAML==3.12
rsa==3.4.2
sendgrid==5.2.0
six==1.10.0
suds-jurko==0.6
tablib==0.12.1
unicodecsv==0.14.1
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
